perform virtualisation
